pyproject.toml
[project]
name = "atlas-swarm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.94.1",
    "claude-agent-sdk>=0.1.59",
    "crewai>=1.14.1",
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "langchain-core>=1.2.28",
    "langgraph>=1.1.6",
    "mlx>=0.31.1",
    "mlx-lm>=0.31.2",
    "pydantic>=2.11.10",
    "pyyaml>=6.0.3",
    "rich>=14.3.4",
    "sqlite-utils>=3.38",
    "uvicorn>=0.44.0",
]
